Renamed and gutted old frontend facade
