+++ src/eth/blockchain.js
@@ -1,5 +1,5 @@
 import web3 from './web3';
 // This blockchainClientUtils is actually a symlink to solidity/client/blockchainUtils.js
-import blockchainUtils from './blockchainClientUtils';
+import blockchainUtils from './blockchainFacade';
 
 export default blockchainUtils(web3);
