--- src/eth/blockchainClientUtils.js
@@ -1,1 +1,0 @@
-../../solidity/client/blockchainUtils.js
