+++ src/eth/blockchainFacade.js
@@ -1,0 +1,1 @@
+../../solidity/facades/frontend.js
